requirements.txt
aioredis==1.2.0
asgiref==3.1.4
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
Beaker==1.10.1
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
colorama==0.4.1
configparser==3.7.4
constantly==15.1.0
daphne==2.2.2
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.9.4
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.20
kiwisolver==1.0.1
lazy-object-proxy==1.4.1
Mako==1.0.13
MarkupSafe==1.1.1
matplotlib==2.2.2
mccabe==0.6.1
mixer==6.1.3
more-itertools==7.0.0
msgpack==0.6.1
mysqlclient==1.4.2.post1
numpy==1.15.0+mkl
oauthlib==3.0.1
oic==1.0.1
olefile==0.45.1
packaging==19.0
Pillow==5.2.0
pluggy==0.12.0
priority==1.3.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pycryptodomex==3.8.2
pyflakes==2.1.1
PyHamcrest==1.9.0
pyjwkest==1.4.2
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pyOpenSSL==19.0.0
pyparsing==2.2.0
pypiwin32==223
pyreadline==2.1
pytest==4.6.3
pytest-cov==2.7.1
pytest-django==3.5.0
python-dateutil==2.7.3
python-quilt==1.0
python3-openid==3.1.0
pytz==2018.4
pywin32==224
requests==2.22.0
requests-oauthlib==1.2.0
scipy==1.1.0
selenium==3.141.0
service-identity==18.1.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
text-unidecode==1.2
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.3.5
typing-extensions==3.7.4
urllib3==1.25.3
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webdriver-manager==1.7
websocket-client==0.56.0
whitenoise==4.1.3
wrapt==1.11.1
zipp==0.5.1
Twisted==19.2.1
txaio==18.8.1
zope.interface==4.6.0
Jinja2==2.10.1 
Werkzeug==0.15.5 
click==7.0 
flask==1.1.1 
itsdangerous==1.1.0